feat(recovery-flyout): set up step handler
