+++ packages/blockchain-wallet-v4-frontend/src/modals/RecoveryPhrase/ShowRecoveryWords/index.tsx
@@ -1,0 +1,8 @@
+import FirstSetWords from './template'
+import React from 'react'
+
+const ShowRecoveryWords = () => {
+  return <FirstSetWords />
+}
+
+export default ShowRecoveryWords
