+++ packages/blockchain-wallet-v4-frontend/src/modals/RecoveryPhrase/index.tsx
@@ -1,13 +1,14 @@
 import { actions, selectors } from 'data'
-import { bindActionCreators, compose } from 'redux'
+import { bindActionCreators, compose, Dispatch } from 'redux'
 
 import { connect } from 'react-redux'
 import { path } from 'ramda'
-// import FirstSetWords from './FirstSetWords'
+import { RootState } from 'data/rootReducer'
 import Flyout, { duration, FlyoutChild } from 'components/Flyout'
 import modalEnhancer from 'providers/ModalEnhancer'
-import React from 'react'
-// import RecoveryPhraseIntroContainer from './RecoveryPhraseIntro'
+import React, { PureComponent } from 'react'
+import RecoveryPhraseIntro from './RecoveryPhraseIntro'
+import ShowRecoveryWords from './ShowRecoveryWords'
 
 export type OwnPropsType = {
   close: () => void
@@ -17,21 +18,25 @@ export type OwnPropsType = {
   userClickedOutside: boolean
 }
 
-type Props = OwnPropsType
-
-type State = {
-  // step:
-  // | 'RECOVERY_PHRASE_INTRO'
-  // | 'FIRST_SET_WORDS'
-  // | 'SECOND_SET_WORDS'
-  // | 'CONFIRM_WORDS'
-  // | 'CONFIRM_WORDS_SUCCESS';
-  show: boolean
+export type LinkDispatchPropType = {
+  recoveryPhraseActions: typeof actions.components.recoveryPhrase
+  settingsActions: typeof actions.modules.settings
 }
 
-class RecoveryPhraseFlyout extends React.PureComponent<Props, State> {
+type LinkStatePropsType = {
+  step:
+    | 'RECOVERY_PHRASE_INTRO'
+    | 'FIRST_SET_WORDS'
+    | 'SECOND_SET_WORDS'
+    | 'CONFIRM_WORDS'
+    | 'CONFIRM_WORDS_SUCCESS'
+}
+
+type Props = OwnPropsType & LinkDispatchPropType & LinkStatePropsType
+type State = { show: boolean }
+
+class RecoveryPhraseFlyout extends PureComponent<Props, State> {
   state: State = {
-    // step: 'RECOVERY_PHRASE_INTRO',
     show: false
   }
 
@@ -39,6 +44,10 @@ class RecoveryPhraseFlyout extends React.PureComponent<Props, State> {
     /* eslint-disable */
     this.setState({ show: true })
     /* eslint-enable */
+  }
+
+  componentWillUnmount () {
+    this.props.recoveryPhraseActions.setStep('RECOVERY_PHRASE_INTRO')
   }
 
   handleClose = () => {
@@ -58,19 +67,36 @@ class RecoveryPhraseFlyout extends React.PureComponent<Props, State> {
         userClickedOutside={userClickedOutside}
         data-e2e='recoveryPhraseModal'
       >
-        <FlyoutChild />
+        {this.props.step === 'RECOVERY_PHRASE_INTRO' && (
+          <FlyoutChild>
+            <RecoveryPhraseIntro
+              {...this.props}
+              handleClose={this.handleClose}
+            />
+          </FlyoutChild>
+        )}
+        {this.props.step === 'FIRST_SET_WORDS' && (
+          <FlyoutChild>
+            <ShowRecoveryWords />
+          </FlyoutChild>
+        )}
       </Flyout>
     )
   }
 }
 
-const mapStateToProps = state => ({
+const mapStateToProps = (state: RootState) => ({
   isMnemonicVerified: selectors.core.wallet.isMnemonicVerified(state),
-  recoveryPhrase: path(['securityCenter', 'recovery_phrase'], state)
+  recoveryPhrase: path(['securityCenter', 'recovery_phrase'], state),
+  step: selectors.components.recoveryPhrase.getStep(state)
 })
 
-const mapDispatchToProps = dispatch => ({
-  settingsActions: bindActionCreators(actions.modules.settings, dispatch)
+const mapDispatchToProps = (dispatch: Dispatch): LinkDispatchPropType => ({
+  settingsActions: bindActionCreators(actions.modules.settings, dispatch),
+  recoveryPhraseActions: bindActionCreators(
+    actions.components.recoveryPhrase,
+    dispatch
+  )
 })
 
 const enhance = compose<any>(
